+++ docs/modules/PageObject.rst
@@ -11,14 +11,7 @@ The PageObject Class
     :undoc-members:
     :show-inheritance:
 
-.. autoclass:: pypdf._utils.ImageFile
+.. autoclass:: pypdf._page.ImageFile
     :members:
+    :inherited-members: File
     :undoc-members:
-    :show-inheritance:
-    :exclude-members: IndirectObject
-
-.. autoclass:: pypdf._utils.File
-    :members:
-    :undoc-members:
-    :show-inheritance:
-    :exclude-members: IndirectObject
